Add string comprehension generator `~[;;]`
