+++ nodes/iternodes.py
@@ -659,6 +659,9 @@ class ComprehensionGen(Expression):
                     self.doElem(subCtx)
             inod.step()
 
+    def fin(self):
+        pass
+
     def do(self, ctx:Context):
         # dprint('ListComprExpr.do0')
         self.res = None # reset prev
@@ -666,6 +669,7 @@ class ComprehensionGen(Expression):
         if len(self.iterNodes) == 0:
             return
         self.iterLoop(0, ctx)
+        self.fin()
 
     def get(self):
         return self.res
@@ -728,12 +732,12 @@ class DictComprExpr(ComprehensionGen):
 
 class BytesComprExpr(ComprehensionGen):
     '''
-        0x[k:v ; k, v <- src; assign? ; cond]
+        0x[n ; n <- src; assign? ; cond]
     '''
     def __init__(self):
         super().__init__()
-        self.resExpr:ServPairExpr
-        self.res:DictVal = None
+        self.resExpr:Expression
+        self.res:BytesVal = None
 
     def resultObject(self):
         ''' '''
@@ -744,4 +748,31 @@ class BytesComprExpr(ComprehensionGen):
         res = self.resExpr.get()
         # print('B-COMPRH . doElem. rexpr:', self.resExpr, 'res:', res, 'val:', var2val(res))
         self.res.addVal(var2val(res))
+
+
+class StringComprExpr(ComprehensionGen):
+    '''
+        ~[s ; s <- src; assign? ; cond]
+    '''
+    def __init__(self):
+        super().__init__()
+        self.resExpr:Expression
+        self.vals = []
+        self.res:StringVal = None
+
+    def resultObject(self):
+        ''' '''
+        self.res = None
+    
+    def fin(self):
+        self.res = StringVal(''.join(self.vals))
+    
+    def doElem(self, ctx:Context):
+        self.resExpr.do(ctx)
+        res = self.resExpr.get()
+        rv = var2val(res).getVal()
+        # print('B-COMPRH . doElem. rexpr:', self.resExpr, 'res:', res, 'val:', rv)
+        if isinstance(rv, Glif):
+            rv = rv.val
+        self.vals.append(rv)
