+++ tests/test_control.py
@@ -30,143 +30,6 @@ from tests.utils import *
 class TestControl(TestCase):
 
 
-    def test_for_multisource_gen(self):
-        ''' test fix for iter-gen and list-expression in for-loop '''
-        code = r'''
-        res = []
-        
-        
-        # iter-gen, list-expr
-        
-        r1 = []
-        # aa = [11,22,33,44]
-        for a, b <- [101..104], [11,22,33,44]
-            r1 <- (a,b)
-        res <- r1
-        
-        # index-gen, iter-gen
-        r2 = []
-        for a, b <- iter(5), [21..25]
-            r2 <- (a, b)
-        res <- r2
-        
-        # iter-gen, list-get
-        r3 = []
-        for a, b <- [1..5], [x ; x <- [11..15]]
-            r3 <- (a, b)
-        res <- r3
-        
-        # function results
-        func foo(n)
-            [1 .. n]
-        
-        func bar(s, sep)
-            s.split(sep)
-        
-        r4 = []
-        for a, b <- foo(3), bar('aa, bb, cc,', ' ')
-            r4 <- (a, b)
-        res <- r4
-        
-        # print('res = ', res)
-        '''
-        code = norm(code[1:])
-        ex = tryParse(code)
-        rCtx = rootContext()
-        ctx = rCtx.moduleContext()
-        trydo(ex, ctx)
-        rvar = ctx.get('res').get()
-        resv = resRepr(rvar.vals())
-        # print(resv)
-        exv = [
-            [(101, 11), (102, 22), (103, 33), (104, 44)], 
-            [(0, 21), (1, 22), (2, 23), (3, 24), (4, 25)], 
-            [(1, 11), (2, 12), (3, 13), (4, 14), (5, 15)],
-            [(1, 'aa,'), (2, 'bb,'), (3, 'cc,')]]
-        self.assertEqual(exv, resv)
-        
-    def test_for_loop_multisource(self):
-        ''' Iterating by several collection sources.
-            for a, b, c <- s1, s2, s3  '''
-        code = r'''
-        res = []
-        
-        # 2 lists
-        nn = [1,2,3,4,5]
-        mm = [10, 20, 30, 40, 50]
-        
-        rr = []
-        for x, y <- nn, mm
-            rr <- (~'{x}+{y}', x + y)
-        res <- rr
-        
-        # 3 lists
-        
-        aa = [1,2,3]
-        bb = [4,5,6]
-        cc = [7,8,9]
-        r2 = []
-        for x, y, z <- aa, bb, cc
-            r2 <- (x, y, z)
-        res <- r2
-        
-        # 5 lists
-        a1 = [1,2,3,4]
-        a2 = [11,22,33,44]
-        a3 = [5,6,7,8]
-        a4 = [55,66,77,88]
-        a5 = [10, 20, 30, 40]
-        r3 = []
-        for q, w, e, r, t <- a1, a2, a3, a4, a5
-            r3 <- (q,w,e,r,t)
-        res <- r3
-        
-        # list, tuple
-        
-        aa = [1,2,3,4,5]
-        tt = ('a','b','c','d','e')
-        r4 = []
-        for a, t <- aa, tt
-            r4 <- (a, t)
-        res <- r4
-        
-        # list, dict
-        
-        aa = [1,2,3,4,5]
-        dd = {'a':'aaa', 'b':'bbb', 'c':'ccc', 'd':'ddd', 'e':'eee'}
-        r5 = []
-        for x, k, v <- aa, dd
-            r5 <- (x, k, v)
-        res <- r5
-        
-        # list, tuple, dict
-        aa = [1,2,3]
-        tt = ('A','B','C')
-        dd = {'aa':111, 'bb':222, 'cc':333}
-        r6 = []
-        for a, t, k, v <- aa, tt, dd
-            r6 <- (a, t, k, v)
-        res <- r6
-        
-        # print('res = ', res)
-        '''
-        code = norm(code[1:])
-        ex = tryParse(code)
-        rCtx = rootContext()
-        ctx = rCtx.moduleContext()
-        trydo(ex, ctx)
-        rvar = ctx.get('res').get()
-        resv = resRepr(rvar.vals())
-        # print(resv)
-        exv = [
-            [('1+10', 11), ('2+20', 22), ('3+30', 33), ('4+40', 44), ('5+50', 55)], 
-            [(1, 4, 7), (2, 5, 8), (3, 6, 9)], 
-            [(1, 11, 5, 55, 10), (2, 22, 6, 66, 20), (3, 33, 7, 77, 30), (4, 44, 8, 88, 40)], 
-            [(1, 'a'), (2, 'b'), (3, 'c'), (4, 'd'), (5, 'e')], 
-            [(1, 'a', 'aaa'), (2, 'b', 'bbb'), (3, 'c', 'ccc'), (4, 'd', 'ddd'), (5, 'e', 'eee')], 
-            [(1, 'A', 'aa', 111), (2, 'B', 'bb', 222), (3, 'C', 'cc', 333)]]
-        self.assertEqual(exv, resv)
-
     def test_match_result(self):
         ''' match-statement return result in the end of function '''
         code = r'''
@@ -451,81 +314,6 @@ class TestControl(TestCase):
         # print(resv)
         exv = [10, 20, 30, 'foo2', 11, 12, 21, 22, 23, 30, 50, 'foo3', 1, 2, 3, 4, 5, 10, 'A.f', 11, 22, 33]
         self.assertEqual(exv, resv)
-
-    def test_for_arrow_multival(self):
-        ''' test arrow-assign `a <- b` 
-            unpack and multiassign if several vars in left: a,b,c <- data
-            no unpack if 1 var: n <- data
-        '''
-        code = r'''
-        res = []
-        
-        # nn = [(1,2,3), (4,5,6)]
-        nnT = [(x, x+2, x+3); i <- [2..5]; x = i * 4]
-        
-        # @debug !123
-        
-        # unpack tuple to vars
-        for a,b,c <- nnT
-            s = ~'{a:02d}:{b:02d}:{c:02d}'
-            res <- s
-        
-        # list of tuples, no unpack
-        for elem <- nnT
-            res <- elem
-        
-        nnL = [[x, x+3, x+5]; i <- [2..5]; x = i * 10]
-        # unpack list to vars
-        for a,b,c <- nnL
-            s = ~'{a:02d}:{b:02d}:{c:02d}'
-            res <- s
-        
-        # list of lists, no unpack
-        for elem <- nnL
-            res <- elem
-        
-        nn2 = [1,2,3,4,5]
-        # iter by list
-        for n <- nn2
-            res <- 'nn2_%d' <<n
-        
-        dd = {1:11, 2:22}
-        
-        # iter by dict, assigning key, val
-        for k, v <- dd
-            res <- ~'dd_{k},{v}'
-        
-        # iter by dict, assigning pair of (key, val)
-        for elem <- dd
-            res <- 'dd_(%d,%d)' << elem
-        
-        # unpack list in generator
-        src = [[1,2], [3,4]]
-        nums = [ x ; sub <- src ; x <- sub]
-        res <- ('nums', nums)
-        
-        # iter by tuple
-        tt = (1,2,3,4,5)
-        for n <- tt
-            res <- ~'tt_{n}'
-        
-        # print('res = ', res)
-        '''
-        code = norm(code[1:])
-        ex = tryParse(code)
-        rCtx = rootContext()
-        ctx = rCtx.moduleContext()
-        trydo(ex, ctx)
-        rvar = ctx.get('res').get()
-        exv = [
-            '08:10:11', '12:14:15', '16:18:19', '20:22:23', 
-            (8, 10, 11), (12, 14, 15), (16, 18, 19), (20, 22, 23), 
-            '20:23:25', '30:33:35', '40:43:45', '50:53:55', 
-            [20, 23, 25], [30, 33, 35], [40, 43, 45], [50, 53, 55], 
-            'nn2_1', 'nn2_2', 'nn2_3', 'nn2_4', 'nn2_5', 
-            'dd_1,11', 'dd_2,22', 'dd_(1,11)', 'dd_(2,22)', 
-            ('nums', [1, 2, 3, 4]), 'tt_1', 'tt_2', 'tt_3', 'tt_4', 'tt_5']
-        self.assertEqual(exv, rvar.vals())
 
     def test_inline_controls(self):
         ''' for /: if /: '''
