+++ tests/test_lists.py
@@ -20,226 +20,6 @@ class TestLists(TestCase):
     ''' Testing lists, iterators, generators, other collections '''
 
 
-    def test_gen_arrow_assign(self):
-        ''' '''
-        code = r'''
-        res = []
-        
-        s = [111]
-        ss = [1,2,3]
-        
-        r1 = [s ; s <- ss]
-        
-        res <- r1
-        res <- s
-        
-        # local context over upper
-        ss = [[1,2], [3,4]]
-        s = [5,6,7]
-        r2 = [(s, n) ; s <- ss; n <- s]
-        res <- 'r2'
-        res <- r2
-        res <- s
-        
-        # for loop
-        s = [1,2,3]
-        ss = [4,5,6]
-        r3 = []
-        for s <- ss
-            r3 <- s
-        res <- 'r3'
-        res <- r3
-        
-        # in dict 
-        
-        s = [111]
-        k = [222]
-        tt = ('qq','ww','ee')
-        ss = [11,22,33]
-        
-        r4 = {k: s ; k, s <- tt, ss}
-        res <- 'r4'
-        res <- r4
-        
-        # `append` huck in gen
-        
-        rt = []
-        func f4(x)
-            rt <- x
-            
-        rp = []
-        f5 = \x -> (rp <- x)
-        
-        s = []
-        ss = [21, 22, 23]
-        
-        r5 = [100 + s ; s <- ss ; _ = f4(s + 200); _= f5(s + 300)]
-        res <- 'r5'
-        res <- r5
-        res <- rt
-        res <- rp
-        
-        # print('res = ', res)
-        '''
-        code = norm(code[1:])
-        ex = tryParse(code)
-        rCtx = rootContext()
-        ctx = rCtx.moduleContext()
-        trydo(ex, ctx)
-        rvar = ctx.get('res').get()
-        resv = resRepr(rvar.vals())
-        # print(resv)
-        exv = [
-            [1, 2, 3], 3, 
-            'r2', [([1, 2], 1), ([1, 2], 2), ([3, 4], 3), ([3, 4], 4)], [3, 4], 
-            'r3', [4, 5, 6], 
-            'r4', {'qq': 11, 'ww': 22, 'ee': 33}, 
-            'r5', [121, 122, 123], [221, 222, 223], [321, 322, 323]]
-        self.assertEqual(exv, resv)
-        
-    def test_dict_generator(self):
-        ''' dict generator
-            { key: val ; key, val <- src; expr; condition }
-        '''
-        code = r'''
-        res = []
-        
-        # simplest case
-        
-        pairs = [('aa', 11), ('bb',22), ('cc',33), ('dd',44)]
-        dd = {k:v ; k, v <- pairs}
-        res <- dd
-        
-        # from list, tuple
-        
-        nn = [1,2,3,4,5]
-        tt = ('a', 'b', 'c', 'd', 'e')
-        
-        d2 = { c : n ; n, c <- nn, tt}
-        d3 = { n : c ; n, c <- nn, tt}
-        
-        res <- d2
-        res <- d3
-        
-        # from slices
-        
-        ss = [1,2,3,4,5, 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', [0]]
-        d4 = { k: v ; v, k <- ss[5:10], ss[:5]}
-        res <- d4
-        
-        # from dict
-        
-        ds = {'a1':'aaa', 'b1':'bbb', 'c1':'ccc', 'd1':'ddd', }
-        d5 = { k:v ; k, v <- ds}
-        res <- d5
-        d6 = { v:k ; k, v <- ds}
-        res <- d6
-        
-        # + condition
-        
-        d7 = { c : n ; n, c <- nn, tt; n > 1 && n % 2 > 0}
-        res <- d7
-        
-        # + pre + condition
-        
-        d8 = { c : s ; n, c <- nn, tt; s = ~'{n:03d}@{c}'; n % 2 > 0}
-        res <- d8
-        
-        # 3 sub levels
-        
-        n9 = [1,2,3]
-        d9 = {
-            k: i * 100 + j ; 
-            n <- n9; p = [1..n]; 
-            i <- p; s = ~'k{i}'; q = [i .. 4]; 
-            j <- q; k = ~'{s}-{j}'; i > 1 && j > 2
-        }
-        res <- d9
-        
-        # unpack list of lists
-        
-        s10 = [
-            [11, 12, 13],
-            [21, 22, 23],
-            [31, 32, 33],
-        ]
-        d10 = {
-            ~'{p}.{n}': n ; 
-            s1 <- s10; 
-            n <- s1; p = s1.fold(n*100, \s, c -> s + c)}
-        res <- d10
-        
-        # print('res = ', res)
-        '''
-        code = norm(code[1:])
-        ex = tryParse(code)
-        rCtx = rootContext()
-        ctx = rCtx.moduleContext()
-        trydo(ex, ctx)
-        # self.assertEqual(0, rvar.getVal())
-        rvar = ctx.get('res').get()
-        resv = resRepr(rvar.vals())
-        # print(resv)
-        exv = [
-            {'aa': 11, 'bb': 22, 'cc': 33, 'dd': 44}, 
-            {'a': 1, 'b': 2, 'c': 3, 'd': 4, 'e': 5}, 
-            {1: 'a', 2: 'b', 3: 'c', 4: 'd', 5: 'e'}, 
-            {1: 'f', 2: 'g', 3: 'h', 4: 'i', 5: 'j'}, 
-            {'a1': 'aaa', 'b1': 'bbb', 'c1': 'ccc', 'd1': 'ddd'}, 
-            {'aaa': 'a1', 'bbb': 'b1', 'ccc': 'c1', 'ddd': 'd1'}, 
-            {'c': 3, 'e': 5}, {'a': '001@a', 'c': '003@c', 'e': '005@e'}, 
-            {'k2-3': 203, 'k2-4': 204, 'k3-3': 303, 'k3-4': 304},
-            {'1136.11': 11, '1236.12': 12, '1336.13': 13, '2166.21': 21, '2266.22': 22, '2366.23': 23, '3196.31': 31, '3296.32': 32, '3396.33': 33}]
-        self.assertEqual(exv, resv)
-
-    def test__listgen_with_multisource(self):
-        '''Multi-source in list-generator
-            [expr ; a, b, c <- s1, s2, s3 ] '''
-        code = r'''
-        res = []
-        
-        # list, tuple
-        aa = [1,2,3]
-        bb = [5,6,7]
-        tt = ('A', 'B', 'C')
-        r1 = []
-        r1 = [{c:(a,b)} ; a, b, c <- aa, bb, tt]
-        res <- r1
-        
-        # iter-gen, dict
-        dd = {'a':'aaa', 'b':'bbb', 'c':'ccc'}
-        r2 = [ ~'{a}:({k},{v})'; a, k, v <- [1 .. len(dd.keys()) ], dd ]
-        res <- r2
-        
-        # index-gen, slice
-        aa = [11, 22, 33, 44, 55, 66, 77]
-        r3 = [(i + 10, x) ; i, x <- iter(3), aa[1:4]]
-        res <- r3
-        
-        # strings
-        s1 = 'RGB'
-        s2 = 'red green blue'
-        r4 = [ ~'{b}:{c}' ; b, c <- list(s1), s2.split(' ') ]
-        res <- r4
-        
-        # print('res = ', res)
-        '''
-        code = norm(code[1:])
-        ex = tryParse(code)
-        rCtx = rootContext()
-        ctx = rCtx.moduleContext()
-        trydo(ex, ctx)
-        # self.assertEqual(0, rvar.getVal())
-        rvar = ctx.get('res').get()
-        resv = resRepr(rvar.vals())
-        # print(resv)
-        exv = [
-            [{'A': (1, 5)}, {'B': (2, 6)}, {'C': (3, 7)}], 
-            ['1:(a,aaa)', '2:(b,bbb)', '3:(c,ccc)'], 
-            [(10, 22), (11, 33), (12, 44)],
-            ['R:red', 'G:green', 'B:blue']]
-        self.assertEqual(exv, resv)
-
     def test_triple_dots_unpack_to_args(self):
         ''' '''
         code = r'''
@@ -701,23 +481,6 @@ class TestLists(TestCase):
         ex.do(ctx)
         rvar = ctx.get('res').get()
         self.assertEqual([1, (2, 3), (4, 5)], rvar.vals())
-
-    def test_list_gen_empty_end(self):
-        ''' list generator. [... expr;] empty last sub-case after semicolon'''
-        code = '''
-        nums1 = [[x ** 2] ; x <- [2..5] ; x > 1; ]
-        nums2 = [[x ** 2 + 1] ; x <- [2..5] ;]
-        res = [nums1, nums2]
-        # print('nums = ', res)
-        '''
-        code = norm(code[1:])
-        ex = tryParse(code)
-        rCtx = rootContext()
-        ctx = rCtx.moduleContext()
-        trydo(ex, ctx)
-        exp = [[[4], [9], [16], [25]], [[5], [10], [17], [26]]]
-        rvar = ctx.get('res').get()
-        self.assertEqual(exp, rvar.vals())
 
     def test_slice_of_generator(self):
         ''' # TODO: nn = [1..10]; nn5 = nn[1:5] '''
